requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
google-generativeai==0.3.2
openai>=2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

